refactor: create `types` file
